refactor(zim): extract _NamespaceMixin to zim/namespace.py
Move namespace listing/browsing/walking methods (list_namespaces,
_list_archive_namespaces, _extract_namespace_from_path,
_get_known_namespace_probes, browse_namespace, _browse_namespace_entries,
_find_entries_in_namespace, _get_common_namespace_patterns,
walk_namespace) into _NamespaceMixin.

ZimOperations now inherits from all four mixins. Drops the now-unused
namespace constants and OpenZimMcpValidationError import from the
zim_operations module.
